requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
soundfile
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
einops==0.8.0
speechbrain==1.0.2

black
isort
pre-commit
